Add StudyReviewListItem
